# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.ProjectCreateReq:
    properties:
      description:
        type: string
      image:
        type: string
      is_visible:
        type: boolean
      project_category_id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        enum:
        - github
        - custom
        type: string
      url:
        type: string
    required:
    - title
    - type
    type: object
...
  internal_handlers.projectVisibilityReq:
    properties:
      is_visible:
        type: boolean
    type: object
...
  portfolio-website-backend_internal_models.JSONMap:
    additionalProperties: true
    type: object
  portfolio-website-backend_internal_models.Project:
    properties:
      additional_data:
        allOf:
        - $ref: '#/definitions/portfolio-website-backend_internal_models.JSONMap'
        description: Store complete GitHub API response
      created_at:
        type: string
      description:
        type: string
      expiry_date:
        description: Expiry date for non-custom projects
        type: string
      id:
        type: string
      image:
        description: Base64 image data
        type: string
      is_visible:
        type: boolean
      project_category:
        $ref: '#/definitions/portfolio-website-backend_internal_models.ProjectCategory'
      project_category_id:
        type: string
      published_at:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        description: e.g. 'github', 'custom'
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  portfolio-website-backend_internal_models.ProjectCategory:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_visible:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is the backend API for my portfolio website.
  title: Portfolio Website API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get a paginated list of projects (requires auth)
      parameters:
      - description: Skip count
        in: query
        name: skip
        type: integer
      - description: Limit count
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project entry
      parameters:
      - description: Project Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ProjectCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/portfolio-website-backend_internal_models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Delete a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get a single project by its ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolio-website-backend_internal_models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update an existing project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Update Data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolio-website-backend_internal_models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /projects/{project_id}/visibility:
    patch:
      consumes:
      - application/json
      description: Change the visibility of a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Visibility Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.projectVisibilityReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolio-website-backend_internal_models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update project visibility
      tags:
      - projects
  /projects/public/{user_id}:
    get:
      consumes:
      - application/json
      description: Get a paginated list of public projects
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Skip count
        in: query
        name: skip
        type: integer
      - description: Limit count
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get public projects
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"